+++ packages/jira-adapter/src/change_validators/index.ts
@@ -40,7 +40,7 @@ import { wrongUserPermissionSchemeValidator } from './wrong_user_permission_sche
 import { GetIdMapFunc } from '../users_map'
 import { accountIdValidator } from './account_id'
 import { screenSchemeDefaultValidator } from './screen_scheme_default'
-import { workflowSchemeDupsValidator } from './workflow_scheme_dups'
+import { workflowSchemeDupsValidator } from './workflows/workflow_scheme_dups'
 import { issueTypeSchemeDefaultTypeValidator } from './issue_type_scheme_default_type'
 import { emptyValidatorWorkflowChangeValidator } from './workflows/empty_validator_workflow'
